backend/requirements.txt
Flask==3.1.*
flask-cors==5.*
psycopg2-binary==2.9.*
python-dotenv==1.1.*
gunicorn==23.*
boto3==1.34.*
PyJWT==2.10.*
